Ejercicio 2 terminado Sistema de empleados y departamento

+++ src/main/kotlin/Ejercicio2/SistemaDeEmpleadosyDepartamentos/EmpleadoFijo.kt
@@ -1,6 +1,6 @@
 package org.example.Ejercicio2.SistemaDeEmpleadosyDepartamentos
 
-class EmpleadoFijo(var salarioAnual: Double, var numeroDePagas: Double, nombre: String): Empleado(nombre) {
+class EmpleadoFijo(var salarioAnual: Double, var numeroDePagas: Int, nombre: String): Empleado(nombre) {
 
     override fun calcularSalario(): Double {
         var salario = salarioAnual / numeroDePagas
